2026.03.19
postArchive 클래스 삭제 - 자주 사용하지 않거나 장기 보존이 필요한 데이터를 운영 시스템에서 별도 저장소로 옮겨 저장하는 기술인데 이 DB는 보통은 Archive log를 이용해서 복구용으로 쓴다고 해서 좀더 공부가 필요할꺼 같아서 현재 DB에서 삭제 요청 후 30일이 지난 데이터를 스케줄을 통해 삭제 하도록 변경함.
PostDeleteBatch 클래스 - Hard Delete 을 위한 batch 클래스
PostResponseDto - viewCount, likeCount의 변수 타입 수정
PostSearchResponseDto - id의 변수 타입 수정
Post 클래스 - 좋아요 수 증가, 감소, 조회수 증가, 감소 비즈니스 메소드 삭제 - 데이터 중복 증가
PostEvent 클래스 - @Bulider 추가, postIds(게시글 목록) 추가
PostEventConsumer - deletePostBulk 메소드 생성 (30일이 지난 후 데이터를 한번에 처리하기 위한 메소드, HardDelete)
PostEventProducer - Topic 상수화, kafkaTemplate 타입 변경, sendPostDeleteBatchEvent 메소드 추가, 이전 코드 수정
PostLikeProducer,ViewCountProducer - Topic 상수화
PostRepository - findDeletePostIdBefore 메소드 추가 및 쿼리문 생성
PostCommandService - Topic 상수화, Redis 캐시 삭제 순서 변경 (구조가 DB 변경 -> 캐시 삭제 -> kafka 이벤트 생성 순으로 되어야함, 즉시 일관성을 위해서 캐시 삭제가 DB 변경 이후에 바로 이루어져야함)
PostLikeService - Topic 상수화, 회원 확인 메소드 삭제 -> 로그인 사용자 기반으로 요청을 컨트롤러에서 하기 때문에 필요가 없다.

+++ src/main/java/com/example/NoticeBoard/domain/post/event/PostEventConsumer.java
@@ -1,14 +1,17 @@
 package com.example.NoticeBoard.domain.post.event;
 
 import com.example.NoticeBoard.domain.post.entity.Post;
+import com.example.NoticeBoard.domain.post.entity.PostEvent;
 import com.example.NoticeBoard.domain.post.entity.PostSearchDocument;
 import com.example.NoticeBoard.domain.post.repository.PostRepository;
 import com.example.NoticeBoard.domain.post.repository.PostSearchRepository;
+import com.example.NoticeBoard.global.enumeration.PostStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 @Component
 @RequiredArgsConstructor
@@ -18,20 +21,21 @@ public class PostEventConsumer {
     private final PostRepository postRepository;
     private final PostSearchRepository postSearchRepository;
     
-    @KafkaListener(topics = "post-event-topic", groupId = "post-group")
-    public void consume(ConsumerRecord<String, Long> record){
-        String eventType = record.key();
-        Long postId = record.value();
+    @KafkaListener(topics = "post-event-topic", groupId = "post-group", concurrency = "3")
+    public void consume(PostEvent event){
+
+        log.info("kafka 이벤트 수신 type={}", event.getEventType());
         
-        log.info("kafka 이벤트 수신 type={}, postId={}", eventType, postId);
-        
-        switch (eventType){
+        switch (event.getEventType()){
             case "CREATE":
             case "UPDATE":
-                indexPost(postId);
+                indexPost(event.getPostId());
                 break;
             case "DELETE":
-                deletePost(postId);
+                deletePost(event.getPostId());
+                break;
+            case "DELETE_BATCH":
+                deletePostBulk(event.getPostIds());
                 break;
         }
     }
@@ -46,9 +50,20 @@ public class PostEventConsumer {
     }
 
     private void deletePost(Long postId) {
-        postSearchRepository.deleteById(postId);
-        log.info("Elasticsearch 삭제 완료 postId={}", postId);
+        Post post = postRepository.findById(postId)
+                .orElseThrow(()-> new RuntimeException("게시글이 존재하지 않습니다."));
+
+        if (post.getPostStatus() == PostStatus.DELETED){
+            postSearchRepository.deleteById(postId);
+            log.info("Elasticsearch 삭제 완료 postId={}", postId);
+        }
     }
 
-
+    // Scheduler가 실행되면(새벽 3시로 설정) 30일이 지난 DELETED 데이터를 조회하면 여러 개의 데이터가 나오는데
+    // DB에서 이걸 전부 Hard DELETE을 하고 Kafka 로 DELETE_BATCH 이벤트를 생성해서
+    // Topic에 전달하고 Consumer에서 처리할 때 호출
+    private void deletePostBulk(List<Long> postIds) {
+        postSearchRepository.deleteAllById(postIds);
+        log.info("Elasticsearch bulk 삭제 완료 count={}", postIds.size());
+    }
 }
